refactor: util 파일에 있는 함수를 쪼갬
